refactor command group handling
