+++ src-tauri/src/restic/command.rs
@@ -5,136 +5,17 @@ use std::{
     fs,
     path::PathBuf,
     process::{Command, Output, Stdio},
-    sync::RwLock,
 };
 
-use lazy_static::lazy_static;
 use scopeguard::defer;
 
 use crate::restic::*;
 
 // -------------------------------------------------------------------------------------------------
 
-#[cfg(target_os = "windows")]
-pub static RESTIC_EXECTUABLE_NAME: &str = "restic.exe";
-#[cfg(target_os = "windows")]
-#[allow(dead_code)]
-pub static RCLONE_EXECTUABLE_NAME: &str = "rclone.exe";
-
-#[cfg(not(target_os = "windows"))]
-pub static RESTIC_EXECTUABLE_NAME: &str = "restic";
-#[cfg(not(target_os = "windows"))]
-#[allow(dead_code)]
-pub static RCLONE_EXECTUABLE_NAME: &str = "rclone";
-
-/// Exit code a process gets killed with via kill_process_with_id.
-#[cfg(target_os = "windows")]
-const COMMAND_TERMINATED_EXIT_CODE: u32 = 288;
-
-// -------------------------------------------------------------------------------------------------
-
-/// Tries to gracefully terminate a process with the provided process ID.
-#[cfg(target_os = "windows")]
-fn terminate_process_with_id(pid: u32) -> Result<(), String> {
-    use windows_sys::Win32::{
-        Foundation::{CloseHandle, GetLastError, BOOL, FALSE, HANDLE, WIN32_ERROR},
-        System::Threading::{OpenProcess, TerminateProcess, PROCESS_TERMINATE},
-    };
-    log::info!("Killing process with PID {}", pid);
-
-    unsafe {
-        // Open the process handle with intent to terminate
-        let handle: HANDLE = OpenProcess(PROCESS_TERMINATE, FALSE, pid);
-        if handle == 0 {
-            let error: WIN32_ERROR = GetLastError();
-            return Err(format!(
-                "Failed to obtain handle to process {}: {:#x}",
-                pid, error
-            ));
-        }
-        // Terminate the process
-        let result: BOOL = TerminateProcess(handle, COMMAND_TERMINATED_EXIT_CODE);
-        // Close the handle now that its no longer needed
-        CloseHandle(handle);
-        if result == FALSE {
-            let error: WIN32_ERROR = GetLastError();
-            return Err(format!("Failed to terminate process {}: {:#x}", pid, error));
-        }
-    }
-    Ok(())
-}
-
-#[cfg(not(target_os = "windows"))]
-fn terminate_process_with_id(pid: u32) -> Result<(), String> {
-    use nix::{
-        sys::signal::{self, Signal},
-        unistd::Pid,
-    };
-    log::info!("Killing process with PID {}", pid);
-    signal::kill(Pid::from_raw(pid as i32), Signal::SIGTERM).map_err(|err| err.to_string())
-}
-
-// -------------------------------------------------------------------------------------------------
-
-lazy_static! {
-    /// Currently running restic command processes mapped by command group names.
-    static ref RUNNING_RESTIC_COMMANDS: RwLock<HashMap<String, Vec<u32>>> =
-        RwLock::new(HashMap::new());
-}
-
-/// Kill all running commands from the given command group.
-fn terminate_all_commands_in_group(command_group: &str) -> Result<(), String> {
-    let running_child_ids = {
-        if let Some(child_ids) = RUNNING_RESTIC_COMMANDS
-            .write()
-            .map_err(|err| err.to_string())?
-            .get_mut(command_group)
-        {
-            std::mem::take(child_ids)
-        } else {
-            vec![]
-        }
-    };
-    if !running_child_ids.is_empty() {
-        log::debug!(
-            "Terminating {} processes in group '{}'...",
-            running_child_ids.len(),
-            command_group
-        );
-        for child_id in running_child_ids {
-            if let Err(err) = terminate_process_with_id(child_id) {
-                log::warn!("Failed to kill command with PID {}: {}", child_id, err);
-            }
-        }
-    }
-    Ok(())
-}
-
-/// Register the given child it with a command group.
-fn add_command_to_group(command_group: &str, child_id: u32) -> Result<(), String> {
-    log::debug!("Process in group '{command_group}' with PID '{child_id}' started...");
-    let mut running_child_ids = RUNNING_RESTIC_COMMANDS
-        .write()
-        .map_err(|err| err.to_string())?;
-    if let Some(child_ids) = running_child_ids.get_mut(command_group) {
-        child_ids.push(child_id);
-    } else {
-        running_child_ids.insert(command_group.to_string(), vec![child_id]);
-    }
-    Ok(())
-}
-
-// Unregister the given child from a command group.
-fn remove_command_from_group(command_group: &str, child_id: u32) -> Result<(), String> {
-    log::debug!("Process in group '{command_group}' with PID '{child_id}' finished");
-    let mut running_child_ids = RUNNING_RESTIC_COMMANDS
-        .write()
-        .map_err(|err| err.to_string())?;
-    if let Some(commands) = running_child_ids.get_mut(command_group) {
-        commands.retain(|id| *id != child_id);
-    }
-    Ok(())
-}
+/// Command group handling
+mod group;
+use group::*;
 
 // -------------------------------------------------------------------------------------------------
 
@@ -152,22 +33,51 @@ pub fn new_command(program: &PathBuf) -> Command {
 
 // -------------------------------------------------------------------------------------------------
 
+#[cfg(target_os = "windows")]
+pub const RESTIC_EXECTUABLE_NAME: &str = "restic.exe";
+#[cfg(target_os = "windows")]
+#[allow(dead_code)]
+pub const RCLONE_EXECTUABLE_NAME: &str = "rclone.exe";
+
+#[cfg(not(target_os = "windows"))]
+pub const RESTIC_EXECTUABLE_NAME: &str = "restic";
+#[cfg(not(target_os = "windows"))]
+#[allow(dead_code)]
+pub const RCLONE_EXECTUABLE_NAME: &str = "rclone";
+
+// -------------------------------------------------------------------------------------------------
+
 /// Restic command executable wrapper.
 #[derive(Debug, Default, Clone)]
 pub struct Program {
-    pub version: [i32; 3],            // restic version [major, minor, rev]
-    pub restic_path: PathBuf,         // path to the restic executable
-    pub rclone_path: Option<PathBuf>, // optional path to rclone executable
+    restic_version: [i32; 3],     // restic version [major, minor, rev]
+    restic_path: PathBuf,         // path to the restic executable
+    rclone_path: Option<PathBuf>, // optional path to rclone executable
 }
 
 impl Program {
-    /// Create a new Restic program with the given path.
+    /// Create a new Restic program with the given path and optional path to rclone.
     pub fn new(restic_path: PathBuf, rclone_path: Option<PathBuf>) -> Self {
         Self {
-            version: Self::version(&restic_path),
+            restic_version: Self::query_restic_version(&restic_path),
             restic_path,
             rclone_path,
         }
+    }
+
+    /// Restic program's versiion (major, minor, rev).
+    pub fn restic_version(&self) -> [i32; 3] {
+        self.restic_version
+    }
+
+    /// Path to the restic executable.
+    pub fn restic_path(&self) -> &PathBuf {
+        &self.restic_path
+    }
+
+    /// Optional path to the rclone executable.
+    pub fn rclone_path(&self) -> &Option<PathBuf> {
+        &self.rclone_path
     }
 
     /// Run a restic command for the given location with the given args.
@@ -273,73 +183,6 @@ impl Program {
         }
     }
 
-    /// Log and return error from a restic run command.
-    fn handle_run_error<S: AsRef<OsStr> + std::fmt::Debug>(args: Vec<S>, output: Output) -> String {
-        // guess if this is a command which got aborted
-        #[cfg(target_os = "windows")]
-        if output
-            .status
-            .code()
-            .is_some_and(|code| code as u32 == COMMAND_TERMINATED_EXIT_CODE)
-        {
-            log::info!("Restic '{:?}' command got aborted", args);
-            return "Command got aborted".to_string();
-        }
-        #[cfg(not(target_os = "windows"))]
-        {
-            use nix::libc::SIGTERM;
-            use std::os::unix::process::ExitStatusExt;
-
-            if output
-                .status
-                .signal()
-                .is_some_and(|status| status == SIGTERM)
-            {
-                log::info!("Restic '{:?}' command got aborted", args);
-                return "Command got aborted".to_string();
-            }
-        }
-        // else log and return stderr as it is
-        let stderr = std::str::from_utf8(&output.stderr).unwrap_or("");
-        log::warn!(
-            "Restic '{:?}' command failed with status {}:\n{}",
-            args,
-            output.status,
-            stderr
-        );
-        stderr.to_string()
-    }
-
-    /// Run restic command to query its version number.
-    fn version(path: &PathBuf) -> [i32; 3] {
-        let mut version = [0, 0, 0];
-        if !path.exists() {
-            return version;
-        }
-        // get version from restic binary
-        match new_command(path).arg("version").output() {
-            Ok(output) => {
-                if output.status.success() {
-                    // "restic x.y.z some other info"
-                    let stdout = std::str::from_utf8(&output.stdout).unwrap_or("");
-                    let mut name_and_version = stdout.split(' ');
-                    if let Some(version_str) = name_and_version.nth(1) {
-                        for (v, s) in version.iter_mut().zip(version_str.split('.')) {
-                            *v = s.parse::<i32>().unwrap_or(0);
-                        }
-                    }
-                } else {
-                    let stderr = std::str::from_utf8(&output.stderr).unwrap_or("");
-                    log::warn!("Failed to read version info from restic binary: {}", stderr);
-                }
-            }
-            Err(err) => {
-                log::warn!("Failed to read version info from restic binary: {err}");
-            }
-        }
-        version
-    }
-
     // Create restic specific args for the given base args and location.
     fn args<'a>(&self, args: Vec<&'a str>, location: &Location) -> Vec<Cow<'a, OsStr>> {
         let mut args = args
@@ -381,5 +224,72 @@ impl Program {
             envs.insert(credential.name, credential.value);
         }
         envs
+    }
+
+    /// Log and return error from a restic run command.
+    fn handle_run_error<S: AsRef<OsStr> + std::fmt::Debug>(args: Vec<S>, output: Output) -> String {
+        // guess if this is a command which got aborted
+        #[cfg(target_os = "windows")]
+        if output
+            .status
+            .code()
+            .is_some_and(|code| code as u32 == COMMAND_TERMINATED_EXIT_CODE)
+        {
+            log::info!("Restic '{:?}' command got aborted", args);
+            return "Command got aborted".to_string();
+        }
+        #[cfg(not(target_os = "windows"))]
+        {
+            use nix::libc::SIGTERM;
+            use std::os::unix::process::ExitStatusExt;
+
+            if output
+                .status
+                .signal()
+                .is_some_and(|status| status == SIGTERM)
+            {
+                log::info!("Restic '{:?}' command got aborted", args);
+                return "Command got aborted".to_string();
+            }
+        }
+        // else log and return stderr as it is
+        let stderr = std::str::from_utf8(&output.stderr).unwrap_or("");
+        log::warn!(
+            "Restic '{:?}' command failed with status {}:\n{}",
+            args,
+            output.status,
+            stderr
+        );
+        stderr.to_string()
+    }
+
+    /// Run restic command to query its version number.
+    fn query_restic_version(path: &PathBuf) -> [i32; 3] {
+        let mut version = [0, 0, 0];
+        if !path.exists() {
+            return version;
+        }
+        // get version from restic binary
+        match new_command(path).arg("version").output() {
+            Ok(output) => {
+                if output.status.success() {
+                    // "restic x.y.z some other info"
+                    let stdout = std::str::from_utf8(&output.stdout).unwrap_or("");
+                    let mut name_and_version = stdout.split(' ');
+                    if let Some(version_str) = name_and_version.nth(1) {
+                        for (v, s) in version.iter_mut().zip(version_str.split('.')) {
+                            *v = s.parse::<i32>().unwrap_or(0);
+                        }
+                    }
+                } else {
+                    let stderr = std::str::from_utf8(&output.stderr).unwrap_or("");
+                    log::warn!("Failed to read version info from restic binary: {}", stderr);
+                }
+            }
+            Err(err) => {
+                log::warn!("Failed to read version info from restic binary: {err}");
+            }
+        }
+        version
     }
 }
